backend/requirements.txt
asgiref==3.8.1
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.9
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
idna==3.4
psycopg2-binary==2.9.6
pycodestyle==2.10.0
PyJWT==2.10.1
pyTelegramBotAPI==4.10.0
requests==2.28.2
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.15
